requirements.txt
praw
openai
langchain
langchain-openai
python-dotenv
textblob
streamlit

